make layer check happy
