+++ src/vs/workbench/workbench.desktop.main.ts
@@ -127,20 +127,21 @@ import './contrib/remote/electron-sandbox/remote.contribution.js';
 // Configuration Exporter
 import './contrib/configExporter/electron-sandbox/configurationExportHelper.contribution.js';
 
+// Output View
+import './contrib/output/electron-sandbox/output.contribution.js';
+
 // Terminal
 import './contrib/terminal/electron-sandbox/terminal.contribution.js';
 
 // Themes
 import './contrib/themes/browser/themes.test.contribution.js';
 import './services/themes/electron-sandbox/themes.contribution.js';
-
 // User Data Sync
 import './contrib/userDataSync/electron-sandbox/userDataSync.contribution.js';
 
 // Tags
 import './contrib/tags/electron-sandbox/workspaceTagsService.js';
 import './contrib/tags/electron-sandbox/tags.contribution.js';
-
 // Performance
 import './contrib/performance/electron-sandbox/performance.contribution.js';
 
@@ -171,7 +172,6 @@ import './contrib/remoteTunnel/electron-sandbox/remoteTunnel.contribution.js';
 // Chat
 import './contrib/chat/electron-sandbox/chat.contribution.js';
 import './contrib/inlineChat/electron-sandbox/inlineChat.contribution.js';
-
 // Encryption
 import './contrib/encryption/electron-sandbox/encryption.contribution.js';
